Refactor into Classes
