+++ data/duke.txt
@@ -1,3 +1,2 @@
 [T][✗] read book
-[D][✗] eat shit on 2019-02-10
 [E][✗] eat shit on 2019-12-01
